Extract other registration items with queued ones
